Standard MPX Global Choropleth

Standard MPX Global Choropleth

Usage

mpx_case_choro(x, breaks = c(1, 2, 6, 20, 99), latest_date = Sys.Date())

Arguments

x

A Monkey pox data.frame with at least iso3code and cases columns

breaks

A numeric vector containing breaks for the choropleth bins

latest_date

A date value for the subtitle, latest updated date (default: Sys.Date())

Value

A styled choropleth map for use in various data products